+++ ts/src/accounts/index.ts
@@ -14,6 +14,14 @@ export {
   type GlobalConfiguration,
 } from "./GlobalConfiguration";
 export {
+  getSpotCollateralMetadataDecoder,
+  getSpotCollateralAssetId,
+  SPOT_COLLATERAL_FLAG_DISABLE_POSITION_AUTHORITY_SWAP,
+  SPOT_COLLATERAL_FLAG_HAS_PERP_ASSET,
+  SPOT_COLLATERAL_FLAG_IS_ACTIVE,
+  type SpotCollateralMetadata,
+} from "./SpotCollateralMetadata";
+export {
   fetchMint,
   decodeMint,
   getMintCodec,